PRODUCT_COPY_FILES: don't copy useless files
This eliminates 'Nvidia Camera Scripting' (nvcs) whose use is
undocumented, along with the entire Python infrastructure needed
to support it. It also removes Broadcom test scripts, dbus support
that was never implemented, and a raft of unused executables and
their supporting libraries.

While these files are no longer being copied into the ROM, they
remain in the repo just-in-case...